backend/requirements.txt
Flask>=2.0
Flask-Cors>=3.0
mysql-connector-python>=8.0
PyJWT>=2.0
bcrypt>=3.2
python-dotenv>=1.0
